*** syncing trunk begins here.
we want to be left with version of CABLE that has ONLY "changes" due to SLI. this does NOT include instances where newer trunk merged version eliminates or modifies line that was untouched in older SLI version. it is i=undoubtedly true that SLI version relied on these in the older format, but we are not going to switch instances where SLI should have updated. 

straight merge trunk 455