moved all computations regarding calculating the equilibrium water content to a seperate subroutine called calc_equilibrium_water_content(ssnow,soil).
testing using the equilibrium water content of the first layer to determine the first layer wetness for evap instead of the field capacity.  thought is that recent rains means evap when first layer is wetter than eq content, but after a while will dry to this value and below will still evap but at a lesser rate.  to do this calc the wtd on first timestep inside calc_srf_wet_fraction.