''Trunk Update'' '''Type:''' Technical code improvement. NB: requires new veg. parameters file. See Comments below. '''Ticket:''' #2
__Overview__
Ticket #2 substitutes hardwiried parameters with TYPEd variables that can be read from an external file/namelist. Also fixes the bug in the old carbon module where CSIRO classification has the number 17 (instead of the original 15) causing a mix-up with the IGBP classification.
__Technical comments__
The code changes in this revision are not included under a switch and furthermore require that a modified version of the vegetation parameters file (def_veg_params.txt) is used (def_veg_params_Ticket2.txt). This file is available in the CABLE-AUX branch in the repository. The cable.nml file will need to be edited to point to this file, or a sym link established.
__Science comments__
neutral. However, UM(7.3) application in an ACCESS1.3b (AMIP) run shows a very minor difference (RE: 0.001%). The source of this error has not been identified but is possibly due to Fortran storing numbers differently when read from file into TYPE members, as opposed to hardwired directly to REALs. This only shows up in UM applications which also clobber all variables at compile time as double precision. Single precision (and offline), GSWP2 runs are bitwise reproducible.