''Trunk Update'' Part  II. Transitory commit of large update - see governing Ticket #128 '''Type:''' Modified Files '''Ticket:''' #128
__Overview__
# see Ticket #128

Whilst this commit is prefaced as transitory, it does pass tests previously applied comprehensively with the Shared branch TicketV. However, these tests only extend to the offline applications of the code. For use in ACCESS it is reccommended that for the time being users checkout an earlier revision, prior to -r3815 (but include modifictions made to cable_radiation.F90 at that revision). This revision will work with the ACCESS model, however it is advised to contact cable_help@nf.nci.org.au if you wish to do so. 

__Technical comments__
Many technical improvements have been made to drivers, I/O routines, error trapping, and bug fixes.

__Science comments__
this update is NOT science neutral. Significant effort has been made to isolate significant changes with user switches that can be configured in the namelist file. Remaining and slight differences can be attributed to the numerous bug fixes etc contained within this update.

See Ticket #128 and links therein for extensive documention of this rather large update.