'Trunk Update''
'''Type:''' 
bug fix

'''Ticket:'''
200

__Overview__
Fixed so that 4D (older ALMA format) or 3D (current ALMA format), or a mix of the two, will work. In doing this I found a bug of sorts: the default met file we have in the repo has 3D PSurf, which was one of the met variables that were expected as 4D, which meant that running CABLE offline with this file used the first time step of PSurf for all time steps. So the fix will not give will not give the same results as the trunk if you test with this file, since PSurf now varies with time, as it should.