''Trunk Update''
'''Type:'''
Improvement

'''Ticket:'''
#71

__Overview__
Initialization of soilsnow variables moved before science code @ t=0. 
This has not been ported to UM version yet as JAC is not ready for this at this stage, howeverthere is no problem forseen. The issue mentioned within the ticket. in which ktau is treated differently betweenAMIP and coupled versions is no longer valid.