merge ESM1.5 modifications Ticket 293: https://trac.nci.org.au/trac/cable/ticket/293
Everything in this ci relating to existing files include zero changes - they are simply moved
We have created a distinct JULES and ESM1.5 directories. These should be made as close to each otehr as possible with most of the files moved to a shared/ directory. In the first instance, ESM1.5 is using its ooriginal CASA files, included in the ESM1.5/ directory. In the CABLEfilesFromESM1.5 are a bunch of driver files specific to ESM1.5 and taken from the CMIP6 submissixon.i There are minimal modifications in these files to include new module names to USE etc. The initialisation directory is common across coupled applications at the moment sas it is not implemented in JAC6.1 anyway. When it is similar files will be consolidated.