''Trunk Update''
'''Type:'''
Code Improvement

'''Ticket:'''
#314

__Overview__
Sync w vn of CABLE@trunk vn (~CABLE3) ported to ESM1.5
Again this SHOULD have no repercussions science wise as either we default to the trunk behaviour (especially as any differences from mESM1.5 ar ein the phyiscal model which is likely scrutinized more so on the trunk side applciation) OR where there are differences that appear to be required by the ESM1.5 model, these have been hidden behind a runtime switch. Running GSWP shows again (curiosly) a tiny (9th decimal place) difference in the time-space-averaged photosynthesis, due to the cbl_soilsnow_main chagnes. Otherwise every other number in the log_cable.txt output is identical. However, this is when we reverse the changes that appear here, i.e. we have let them through. The first of the two changes below has led to a very minor difference in the cummulative energy balnce calculated per cell. Some better, some worse. These changes are passed here, however will be reviewed in greater length, as the changes are probably more correct. These changes are in 2 files:

cbl_stempv.F90: soil heat capacity is calculated within a loop over the soil levels (k) AND has k dimensions availabe, however was hard-wired to use the heat capacity of the surface soil layer. This has been modified to use the kth level.  

cbl_soilsnow_main.F90: hidden behind a IF(soil_thermal_fix switch) was clobbering of the heat_capacity to =0.01. So that it never goes to zero. This has been revised so that the heat capacity is calculated everywhere BUT is limited in range to be 0.01 or greater.
There are 2 local variables, xx and xxx. ESM1.5 has oenn defined as r_2, and the other as REAL. The trunk - vice-versa.  Believe it or not this was another source leading to difference in enrgy/water balances (albeit tiny). We default to ther trunk version.