''Trunk Update''
'''Type:'''
Give default value to gs_switch

'''Ticket:'''
#332

__Overview__
CABLE won't run without a value for gs_switch being defined. The intention was to force the user to specify a stomatal conductance scheme, however this is reported as a problem often enough that is preferable to define a default scheme here. This was intended to upload as aprt of #332, however the bulk of #332 has been put on hold until a thourough assesment can be achieved